feature: Adds roles to idm
Changes the idm to have the potential to house a wider range of roles by
providing a json field.

Changes the idm user to have the concept of an 'application' meaning
the idm can be used for multiple applications.

Also adds created and updated dates to the user table for basic
auditing.