Remove config.js tests
Across all 8 apps that make up the service this was the only one to have tests for config.js. They are very much centered in what should happen when `TEST_MODE` is 'truthy' or 'falsey'.

Problem is, the tests don't actually line up with what is in config.js. Yes, depending on `TEST_MODE` the log level would change. But those that talk about base url's being set depending on `TEST_MODE` are no longer valid. Those values have nothing to do with it.

When you take out the references to `TEST_MODE` what's there comes down to testing if an en var is set we use it, else we use a default. And IMHO that is of no value.

So, rather than waste timing trying to make them meaningful we delete the tests and bring this project inline with all our other repos.