Set log level using new env var (#796)
https://github.com/DEFRA/water-abstraction-team/issues/41

We're currently unable to set the log level used within the app. Instead, it's determined by looking at the env var `TEST_MODE`. If 'truthy' it sets the log level to `info`, else the default is `error`.

But we'd actually like to run with `debug` locally and `info` in **production**. And if that works out too noisy perhaps try `warn`. The key thing is we would expect to be able to explicitly set the log level for an app in an environment and currently we can't.

So, this change removes the reliance on `TEST_MODE` for setting the log level and instead will read it directly from an env var (if set).

** Notes - Remove config.js tests

Across all 8 apps that make up the service this was the only one to have tests for config.js. They are very much centred on what should happen when `TEST_MODE` is 'truthy' or 'falsey'.

Problem is, that the tests don't actually line up with what is in config.js. Yes, depending on `TEST_MODE` the log level would change. But those that talk about base URLs being set depending on `TEST_MODE` are no longer valid. Those values have nothing to do with it.

When you take out the references to `TEST_MODE` what's there comes down to testing if an en var is set we use it, else we use a default. And IMHO that is of no value.

So, rather than waste timing trying to make them meaningful we delete the tests and bring this project inline with all our other repos.