Remove unneeded packages (#799)
https://github.com/DEFRA/water-abstraction-team/issues/26

We have some dependencies which aren't needed at all -- for example, they may be things that were originally in use but were since dropped, and have never been removed from package.json. This is slightly different to the issue of packages which are in use but are deprecated so keep the two separate.