Fix external change user email journey
https://eaflood.atlassian.net/browse/WATER-4632

We received a report that the external app's change email journey was broken. We quickly confirmed this.

The problem was this section in `src/modules/change-email/controller.js`.

```javascript
const postStartEmailChange = async (request, h) => {
  const { userId } = request.params
  const { email } = request.payload

  try {
    // Rate limit requests
    const existing = await repos.changeEmailRepo.findOneByUserId(userId)

    validateEmailChange(userId, existing)

    // Upsert email change record
    const data = await repos.changeEmailRepo.create(userId, email)

    // ...
```

The controller was essentially rebuilt in [Bugfix/water 2188 further work](https://github.com/DEFRA/water-abstraction-tactical-idm/pull/281) (Aug 2019) and that's when it was left like this.

The idea of the `validateEmailChange()` call is to ensure that the a retry limit (which they named rate limit) has not been hit, and the user has been re-authenticated (the journey requires you to re-enter your password). It does this using the data returned by `await repos.changeEmailRepo.findOneByUserId(userId)`.

But hold up, when does the record get created that holds the data being validated? That's right, after the validation call! (`const data = await repos.changeEmailRepo.create(userId, email)`)

Any user attempting to change their email will find it blows up because `existing` never contains anything. This means the feature has never worked since the 'bugfix' was implemented. 😱🤦

So, in this change we do 2 things

- we only call `validateEmailChange()` if we have something to validate
- we tweak the `getStatus()` handler to include the security code needed to verify the email change

That second change relates to work we've done to support [Adding an external user change email test](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/131) to our automated test suite. Hopefully if we break it, we'll spot the problem sooner than 5 years. 😜